refactor: 그룹 DTO 패키지를 세부 도메인 기반으로 분리 (#220)
