+++ src/main/java/com/moa/moa_server/domain/group/controller/GroupVoteController.java
@@ -1,0 +1,3 @@
+package com.moa.moa_server.domain.group.controller;
+
+public class GroupVoteController {}
