+++ src/main/java/com/moa/moa_server/domain/group/service/GroupVoteService.java
@@ -1,0 +1,3 @@
+package com.moa.moa_server.domain.group.service;
+
+public class GroupVoteService {}
